move pytential.solve to pytential.linalg.gmres
