--- doc/tools.rst
@@ -1,6 +1,0 @@
-Utilities
-=========
-
-.. automodule:: pytential.solve
-
-.. vim: sw=4:fdm=marker
